ecomm/requirement.txt
apturl==0.5.2
arabic-reshaper==2.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asn1crypto==1.5.1
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.11.1
bleach==3.3.0
blinker==1.4
braintree==4.6.0
Brlapi==0.7.0
certifi==2021.5.30
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.3
command-not-found==0.3
cryptography==38.0.1
cssselect2==0.6.0
cupshelpers==1.0
dbus-python==1.2.16
debugpy==1.6.0
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2
django-allauth==0.39.1
django-countries==5.3.3
django-crispy-forms==1.14.0
django-debug-toolbar==1.10.1
django-filter==22.1
django-environ==0.9.0
django-heroku==0.3.1
django-jet-reboot==1.1.1
django-js-asset==2.0.0
django-mptt==0.13.4
django-paypal==1.1.1
djangorestframework==3.12.4
docx==0.2.4
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.1.0
executing==0.8.3
Faker==13.12.1
fasteners==0.14.1
fastjsonschema==2.15.3
feedparser==6.0.8
filelock==3.7.1
fpdf==1.7.2
fpdf2==2.5.5
future==0.18.2
gunicorn==20.1.0
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
importlib-resources==5.8.0
ipykernel==6.15.0
ipython==8.4.0
ipython_genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.6.1
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.6.3
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==2.6.6
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
monotonic==1.5
more-itertools==8.14.0
mysql-connector-python-rf==2.2.2
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
netifaces==0.10.4
notebook==6.4.12
numpy==1.23.0
oauthlib==3.0.1
olefile==0.46
openpyxl==3.0.10
oscrypto==1.3.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
pathlib==1.0.1
pbr==5.10.0
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==6.2.2
platformdirs==2.5.2
pluggy==0.13.1
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.6.1
psutil==5.9.1
psycopg2==2.8.6
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycairo==1.16.2
pycodestyle==2.5.0
pycparser==2.21
pycups==1.9.73
Pygments==2.12.0
PyGObject==3.36.0
pyHanko==0.13.2
pyhanko-certvalidator==0.19.5
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyparsing==3.0.9
PyPDF3==1.0.6
pyRFC3339==1.1
pyrsistent==0.18.1
pytest==5.4.1
pytest-django==3.9.0
python-apt==2.0.0+ubuntu0.20.4.7
python-bidi==0.4.2
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-decouple==3.1
python-docx==0.8.10
python-dotenv==0.12.0
python3-openid==3.1.0
pytz==2018.5
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.26
PyYAML==5.3.1
pyzmq==23.2.0
qrcode==7.3.1
qtconsole==5.3.1
QtPy==2.1.0
razorpay==1.3.0
reportlab==3.5.67
requests==2.21.0
requests-oauthlib==1.2.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
Send2Trash==1.8.0
sgmllib3k==1.0.0
simplejson==3.16.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.2.4
stack-data==0.3.0
static3==0.7.0
stripe==2.27.0
svglib==1.4.1
systemd-python==234
terminado==0.15.0
testresources==2.0.1
tinycss2==1.1.1
tk==0.1.0
tornado==6.1
tqdm==4.64.1
traitlets==5.3.0
typing-extensions==3.10.0.0
tzdata==2022.2
tzlocal==4.2
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
Unipath==1.0
uritools==4.0.0
urllib3==1.24.2
usb-creator==0.3.7
virtualenv==20.14.1
virtualenvwrapper-win==1.2.7
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
widgetsnbextension==3.6.1
xhtml2pdf==0.2.8
xkit==0.0.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.0
